由鹭青发起的VOD SDK自动发布, 版本号：2.10.0 发布日志： 1, Add the DescribePlayTopVideos  API,  to describe the top video statistics of play data. 2, Add the DescribePlayUserAvg API, to describe the average statistics of play data. 3, Add the DescribePlayUserTotal API, to describe the total statistics of play data. 4, Add the DescribePlayVideoStatis API, to describe the single video statistics of play data. 5, Add the DescribeRefreshQuota API, to get quota of refreshing cache. 6, Add the DescribeRefreshTasks API, to get status of refreshing cache. 7, Add the PushObjectCache API, to summit a job of pushing objects to cdn cache. 8, Add the RefreshObjectCaches API, to refresh cdn cache. 9, Add the SubmitTranscodeJobs API, to support the user to submit transocde job base with VideoId, TemplateGroupId, EncryptConfig and so on. 10, Add the SubmitSnapshotJob API, to support the user to submit snapshot job base with VideoId, SpecifiedOffsetTime, Width, Height and so on. 11, Update the ProduceEditingProjectVideo API, add MediaMetadata field of request to set the produced media metadata, add ProduceConfig field of request to set some customized config options, such as TemplateGroupId and so on.

+++ aliyun-java-sdk-vod/src/main/java/com/aliyuncs/vod/model/v20170321/CreateUploadVideoRequest.java
@@ -1,21 +1,17 @@
 /*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
+
 package com.aliyuncs.vod.model.v20170321;
 
 import com.aliyuncs.RpcAcsRequest;
@@ -56,7 +52,7 @@ public class CreateUploadVideoRequest extends RpcAcsRequest<CreateUploadVideoRes
 
 	private String templateGroupId;
 
-	private Integer cateId;
+	private Long cateId;
 
 	public Long getResourceOwnerId() {
 		return this.resourceOwnerId;
@@ -201,11 +197,11 @@ public class CreateUploadVideoRequest extends RpcAcsRequest<CreateUploadVideoRes
 		}
 	}
 
-	public Integer getCateId() {
+	public Long getCateId() {
 		return this.cateId;
 	}
 
-	public void setCateId(Integer cateId) {
+	public void setCateId(Long cateId) {
 		this.cateId = cateId;
 		if(cateId != null){
 			putQueryParameter("CateId", cateId.toString());
