+++ aliyun-java-sdk-vod/src/main/java/com/aliyuncs/vod/model/v20170321/GetVideoListRequest.java
@@ -1,21 +1,17 @@
 /*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
+
 package com.aliyuncs.vod.model.v20170321;
 
 import com.aliyuncs.RpcAcsRequest;
@@ -34,11 +30,11 @@ public class GetVideoListRequest extends RpcAcsRequest<GetVideoListResponse> {
 
 	private String resourceOwnerAccount;
 
-	private Integer cateId;
+	private Long cateId;
 
-	private Integer pageNo;
+	private Long pageNo;
 
-	private Integer pageSize;
+	private Long pageSize;
 
 	private String endTime;
 
@@ -72,33 +68,33 @@ public class GetVideoListRequest extends RpcAcsRequest<GetVideoListResponse> {
 		}
 	}
 
-	public Integer getCateId() {
+	public Long getCateId() {
 		return this.cateId;
 	}
 
-	public void setCateId(Integer cateId) {
+	public void setCateId(Long cateId) {
 		this.cateId = cateId;
 		if(cateId != null){
 			putQueryParameter("CateId", cateId.toString());
 		}
 	}
 
-	public Integer getPageNo() {
+	public Long getPageNo() {
 		return this.pageNo;
 	}
 
-	public void setPageNo(Integer pageNo) {
+	public void setPageNo(Long pageNo) {
 		this.pageNo = pageNo;
 		if(pageNo != null){
 			putQueryParameter("PageNo", pageNo.toString());
 		}
 	}
 
-	public Integer getPageSize() {
+	public Long getPageSize() {
 		return this.pageSize;
 	}
 
-	public void setPageSize(Integer pageSize) {
+	public void setPageSize(Long pageSize) {
 		this.pageSize = pageSize;
 		if(pageSize != null){
 			putQueryParameter("PageSize", pageSize.toString());
