+++ aliyun-java-sdk-vod/src/main/java/com/aliyuncs/vod/model/v20170321/GetVideoListResponse.java
@@ -1,21 +1,17 @@
 /*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
+
 package com.aliyuncs.vod.model.v20170321;
 
 import java.util.List;
@@ -83,7 +79,7 @@ public class GetVideoListResponse extends AcsResponse {
 
 		private String coverURL;
 
-		private Integer cateId;
+		private Long cateId;
 
 		private String cateName;
 
@@ -177,11 +173,11 @@ public class GetVideoListResponse extends AcsResponse {
 			this.coverURL = coverURL;
 		}
 
-		public Integer getCateId() {
+		public Long getCateId() {
 			return this.cateId;
 		}
 
-		public void setCateId(Integer cateId) {
+		public void setCateId(Long cateId) {
 			this.cateId = cateId;
 		}
